first

+++ src/main/groovy/examples/ThingsDB.groovy
@@ -2,13 +2,6 @@ package examples
 
 import net.dietmaier.groovy.rest.Resterl
 
-void handleETag(Resterl thing ) {
-  def etag = thing.last.headers.ETag
-
-  if ( etag )
-    thing.requestProperties.'If-Match' = etag
-}
-
 // Base-URL, prefix and basicAuth will be kept in the instance
 Resterl things = new Resterl([
           url: 'http://localhost:8080',
@@ -20,6 +13,7 @@ things.query([
     desc: 'a database with very important things'
 ]).put()
 
+// check response
 assert [
   200, // returned if DB already exists
   201  // returned if DB is newly created
@@ -28,23 +22,24 @@ assert [
 // create first collection
 def colored = things.sub('colored')
 
-def curr = colored.get()
-handleETag(colored)
-colored.delete() // delete if exists
-println "delete collection: ${colored.last.code}"
+// get status of colored
+colored.get()
+// delete it of it exists
+if ( colored.last.valid ) {
+  RestheartHelper.setMatch(colored)
+  colored.delete() // delete if exists
+}
 
+//recreate collection
 colored.query([
     desc: 'a collection for colored things'
 ]).put()
 
 // add some things
-
 colored.body([
     name: 'banana',
     color: 'yellow'
 ]).post()
-
-println colored.get()._embedded
 
 colored.body([
     [ name: 'rose', color: 'red'],
@@ -52,31 +47,19 @@ colored.body([
     [ name: 'yellow brick road', color: 'pink' ]
 ]).post()
 
-def current = colored.get()._embedded
-
-current.each {
-  println it
-}
-
 // correct a mistake
-def ybr = current.find{it.name == 'yellow brick road'}
-colored.requestProperties.'If-Match' = ybr._etag.'$oid'
+def ybr = colored.get()._embedded.find { it.name == 'yellow brick road'}
 
-def id = ybr._id.'$oid'
+RestheartHelper.setMatch(colored, ybr)
+def id = RestheartHelper.idOf(ybr)
 ybr.color = 'yellow'
-ybr.remove('_etag')
-ybr.remove('_id')
 
-println "putting: $ybr"
-
+// upsert document - update would need PATCH - Support
 colored.path(id).body(
-    ybr
+    RestheartHelper.sanitized(ybr)
 ).put()
 
-colored.last.dumpparams()
-
-current = colored.get()
-
-current._embedded.each {
-  println it
+// a look at the result
+colored.get()._embedded.each {
+  println RestheartHelper.sanitized(it)
 }
